pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Fourth Hexlet Project"
authors = ["Vladimir Bazhanov <vladimir.a.bazhanov@gmail.com>"]
readme = "README.md"
packages = [
    { include = 'task_manager' }
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.3"
django-extensions = "^3.2.3"
rollbar = "^0.16.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
python-dotenv = "^1.0.0"
coverage = "^7.3.1"
faker = "^19.6.1"
factory-boy = "^3.3.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
